전역변수로 사용할 prodList, lastSel은 store 디렉토리로 이동
