+++ src/basic/logic.js
@@ -1,6 +1,7 @@
-import { lastSel, prodList, updateLastSelValue } from "./main.basic.js";
 import { ElementIds } from "../constants.js";
 import { calcCart } from "./calcCart.js";
+import { prodList } from "./store/prodList.js";
+import { lastSel, updateLastSelValue } from "./store/lastSel.js";
 
 function updateSelOpts(selElem) {
   const sel = selElem ?? document.getElementById(ElementIds.SEL);
@@ -97,21 +98,15 @@ function createContent(wrap) {
   const cont = document.createElement("div");
   cont.className = "bg-gray-100 p-8";
   cont.appendChild(wrap);
-  return cont;
-}
 
-function createWrap(children) {
-  const wrap = document.createElement("div");
-  wrap.className =
-    "max-w-md mx-auto bg-white rounded-xl shadow-md overflow-hidden md:max-w-2xl p-8";
-  children.map((child) => wrap.appendChild(child));
-  return wrap;
+  return cont;
 }
 
 function createHeaderTxt() {
   const hTxt = document.createElement("h1");
   hTxt.className = "text-2xl font-bold mb-4";
   hTxt.textContent = "장바구니";
+
   return hTxt;
 }
 
@@ -155,6 +150,7 @@ function handleClickCartDisp(event) {
       prod.q += remQty;
       itemElem.remove();
     }
+
     calcCart();
   }
 }
@@ -163,6 +159,7 @@ function createCartDisp() {
   const cartDisp = document.createElement("div");
   cartDisp.id = ElementIds.CART_DISP;
   cartDisp.addEventListener("click", handleClickCartDisp);
+
   return cartDisp;
 }
 function createSum() {
@@ -172,6 +169,7 @@ function createSum() {
 
   return sum;
 }
+
 function createSel() {
   const sel = document.createElement("select");
   sel.id = ElementIds.SEL;
@@ -186,14 +184,25 @@ function createAddBtn() {
   addBtn.id = ElementIds.ADD_BTN;
   addBtn.textContent = "추가";
   addBtn.addEventListener("click", handleClickAddBtn);
+
   return addBtn;
 }
+
 function createStockInfo() {
   const stockInfo = document.createElement("div");
   stockInfo.id = ElementIds.STOCK_INFO;
   stockInfo.className = "text-sm text-gray-500 mt-2";
 
   return stockInfo;
+}
+
+function createWrap(children) {
+  const wrap = document.createElement("div");
+  wrap.className =
+    "max-w-md mx-auto bg-white rounded-xl shadow-md overflow-hidden md:max-w-2xl p-8";
+  children.map((child) => wrap.appendChild(child));
+
+  return wrap;
 }
 
 export function createRootChildren() {
