+++ src/basic/main.basic.js
@@ -5,29 +5,11 @@ import {
 } from "./logic.js";
 import { calcCart } from "./calcCart.js";
 
-export let lastSel;
-export function updateLastSelValue(newValue) {
-  lastSel = newValue;
-}
-
-export const prodList = [
-  { id: "p1", name: "상품1", val: 10000, q: 50 },
-  { id: "p2", name: "상품2", val: 20000, q: 30 },
-  { id: "p3", name: "상품3", val: 30000, q: 20 },
-  { id: "p4", name: "상품4", val: 15000, q: 0 },
-  { id: "p5", name: "상품5", val: 25000, q: 10 },
-];
-export function getProductList() {
-  return prodList;
-}
-export function findProduct(id) {
-  return prodList.find((prod) => prod.id === id);
-}
-
 function main() {
   const root = document.getElementById("app");
   const cont = createRootChildren();
   root.appendChild(cont);
+
   calcCart();
   setSaleAlert();
   setSuggestionAlert();
