+++ src/basic/store/lastSel.js
@@ -1,0 +1,5 @@
+export let lastSel;
+
+export function updateLastSelValue(newValue) {
+  lastSel = newValue;
+}
