+++ src/basic/store/prodList.js
@@ -1,0 +1,15 @@
+export const prodList = [
+  { id: "p1", name: "상품1", val: 10000, q: 50 },
+  { id: "p2", name: "상품2", val: 20000, q: 30 },
+  { id: "p3", name: "상품3", val: 30000, q: 20 },
+  { id: "p4", name: "상품4", val: 15000, q: 0 },
+  { id: "p5", name: "상품5", val: 25000, q: 10 },
+];
+
+export function getProductList() {
+  return prodList;
+}
+
+export function findProduct(id) {
+  return prodList.find((prod) => prod.id === id);
+}
